+++ R/modActiveBinding.R
@@ -1,7 +1,6 @@
 makeSimListActiveBindings <- function(sim) {
   mods <- modules(sim)
   mods <- mods[mods %in% ls(sim@.xData$.mods)]
-  #browser()
   lapply(mods, function(mod) {
     makeModActiveBinding(sim = sim, mod = mod)
     makeParActiveBinding(sim = sim, mod = mod)
@@ -20,10 +19,6 @@ makeModActiveBinding <- function(sim, mod) {
                       fun = activeModBindingFunction,
                       env = env)
     }
-
-  if (exists("aaaaa", envir = .GlobalEnv, inherits = FALSE))
-    browser()
-
 
 }
 
@@ -46,8 +41,6 @@ activeModBindingFunction <- function(value) {
   ret <- NULL
   if (missing(value)) {
     simEnv <- try(whereInStack("sim"), silent = TRUE)
-    if (exists("aaaaa", envir = .GlobalEnv, inherits = FALSE))
-      browser()
     if (!is(simEnv, "try-error")) {
       sim <- try(get("sim", simEnv, inherits = FALSE), silent = TRUE)
       if (!is(sim, "try-error")) {
@@ -57,7 +50,6 @@ activeModBindingFunction <- function(value) {
       }
     }
   } else {
-    browser()
     stop("Can't overwrite mod")
   }
   return(ret)