+++ revdep/check.R
@@ -1,7 +1,13 @@
 #remotes::install_github("r-lib/revdepcheck")
 library("revdepcheck")
 
-options(repos = c(CRAN = 'https://cloud.r-project.org'))
+options(repos = c(
+  PE = "https://predictiveecology.r-universe.dev",
+
+  ## note this is counter to the "Canonical CRAN.r-project.org" for CRAN packages,
+  ## but revdep/ directory is .Rbuildignored and this is used for revdep checks only.
+  CRAN = paste0("https://", "cloud.", "r-project.", "org")
+))
 
 revdepcheck::revdep_reset()
 revdepcheck::revdep_check(num_workers = getOption("Ncpus", 8), quiet = FALSE, timeout = 30*60) ## 30 mins